moved model creator to new package
