+++ src/main/java/mouse/hoi/gamefiles/transform/SimpleCollectionManagerImpl.java
@@ -1,7 +1,7 @@
 package mouse.hoi.gamefiles.transform;
 
 import mouse.hoi.gamefiles.tempmodel.common.SimpleCollection;
-import mouse.hoi.gamefiles.parser.ModelCreator;
+import mouse.hoi.gamefiles.common.modelcreator.ModelCreator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
